+++ kernel/buffer.h
@@ -11,7 +11,7 @@ struct buffer {
 
 void buffer_init(struct buffer *ctx, void *buffer, size_t size);
 err_t buffer_write(struct buffer *ctx, const void *buffer, u64 count,
-                   u64 offset, u64 *out);
+                   u64 offset, size_t *out);
 err_t buffer_read(struct buffer *ctx, void *buffer, u64 count, u64 offset,
-                  u64 *out);
+                  size_t *out);
 #endif // BUFFER_H
